+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * [INTERAL] Renamed `Axn::Enqueueable` to `Axn::Async`
 * [BREAKING] Replaced `.enqueue` (only supported sidekiq) with `.call_async` (via a configurable registry of backgrounding libraries)
 * [FEAT] attachable now creates a foo_async to call call_async
+* `type` validator is not still applied to the blank value when allow_blank is true (`type: Hash` will no longer accept `false` or `""`)
 
 ## 0.1.0-alpha.2.8.1
 * [BUGFIX] Fixed symbol callback and message handlers not working in inherited classes due to private method visibility issues